% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_data.R
\docType{data}
\name{deseq_data}
\alias{deseq_data}
\title{deseq2 output}
\format{A data frame}
\source{
\url{http://www.diamondse.info/}
}
\usage{
deseq_data
}
\description{
A dataset print by deseq2
deseq_data
}
\keyword{datasets}
