% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumd.R
\name{epmap}
\alias{epmap}
\title{match exp and pd in order}
\usage{
epmap(exp, pd)
}
\arguments{
\item{exp}{A numeric matrix export by exprs function}

\item{pd}{A meta data.frame export by pData function}
}
\value{
make exp match with pd in order
}
\description{
make exp match with pd in order
}
\examples{
epmap(exp,pd)
}
